Update GTAsust.cpp
creating the body and coping some functions from varies source to improve the game functionalities
the copied snippet code is "CONSOLE_CURSOR_INFO cursorInfo;
    GetConsoleCursorInfo(consoleHandle, &cursorInfo);
    cursorInfo.bVisible = false;
    SetConsoleCursorInfo(consoleHandle, &cursorInfo);"
this code is retrieving the handle to the console output, getting information about the console cursor, modifying the visibility of the cursor to make it invisible, and then applying these changes to the console. This is often done to improve the aesthetics of console-based applications where a blinking cursor might not be desired.